requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
